server: Fix page size calculation in write access check.

+++ server/ptrace.c
@@ -413,7 +413,7 @@ int read_process_memory( struct process *process, client_ptr_t ptr, data_size_t 
 /* len is the total size (in ints) */
 static int check_process_write_access( struct thread *thread, long *addr, data_size_t len )
 {
-    int page = get_page_size() / sizeof(int);
+    int page = get_page_size() / sizeof(long);
 
     for (;;)
     {
